Move regex functions into core and update imports
