--- lib/regex/isValidRegex.ts
@@ -1,9 +1,0 @@
-export const isValidRegex = (pattern: string): boolean => {
-  if (!pattern) return true;
-  try {
-    new RegExp(pattern);
-    return true;
-  } catch {
-    return false;
-  }
-};
